src/rplidar_python/pyproject.toml
[tool.poetry]
name = "rplidar_sdk"
version = "0.1.0"
description = "Python wrappers around the Slamtek C++ rplidar driver"
authors = ["Joshua Salzedo <thhunkn0wnd@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6"
attrs = "*"

[tool.poetry.dev-dependencies]
ipython = "^7.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
